requirements.txt
dtw-python==1.1.10
matplotlib
numba
numpy
pandas
patsy==0.5.1
Pillow==8.2.0
pycosat==0.6.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
statsmodels==0.12.2
threadpoolctl==2.1.0
